optimized some code

+++ include/ck_tile/ops/fmha/kernel/fmha_batch_prefill_kernel.hpp
@@ -709,7 +709,7 @@ struct FmhaBatchPrefillWithPagedKVCacheKernel
         const float sink_value =
             kargs.sink_ptr != nullptr
                 ? (*(static_cast<const float*>(kargs.sink_ptr) + i_nhead)) / kargs.scale_s
-                : static_cast<float>(-numeric<half_t>::infinity());
+                : -numeric<float>::infinity();
         const index_t seqlen_k = [&]() {
             if constexpr(kKVLookupTable ==
                          BlockAttentionKVCacheLookupTableEnum::SGLANG_PAGE_TABLE_1D)
