+++ include/ck_tile/ops/fmha/kernel/fmha_fwd_splitkv_kernel.hpp
@@ -622,7 +622,7 @@ struct FmhaFwdSplitKVKernel
         const float sink_value =
             kargs.sink_ptr != nullptr
                 ? (*(static_cast<const float*>(kargs.sink_ptr) + i_nhead)) / kargs.scale_s
-                : static_cast<float>(-numeric<half_t>::infinity());
+                : -numeric<float>::infinity();
 
         if constexpr(kIsGroupMode)
         {
